Safari-609.1.11.1/LayoutTests/svg/hittest/block-svg-expected.txt

 PASS: element at 28, 20 is svg
PASS: element at 28, 130 is svg
